requirements.txt
absl-py==0.7.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.82
botocore==1.12.82
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.5
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
gast==0.2.2
gensim==3.4.0
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.13
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
mock==2.0.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.4
olefile==0.46
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plac==0.9.6
preshed==2.0.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.7.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.3.1
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.9
pyzmq==17.1.2
qtconsole==4.4.3
regex==2018.11.22
requests==2.21.0
s3transfer==0.1.13
scipy==1.1.0
Send2Trash==1.5.0
six==1.12.0
smart-open==1.8.0
spacy==2.0.16
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.3.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==6.12.1
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.3.1
ujson==1.35
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
